Convert a couple of examples to a form that can be run with py.test
The tests have preserved their original behavior as much as possible, except in
some cases where output was given, assertions are made to match the expected
output.

chapel.inline.numpy.py went to test_chapel_slicing_inline.py, c.sfile.py went
to test_c_sfile.py, chapel.numpy.explicit.py went to test_pych_reduce_add.py
and chapel.inline.reduce.py went to test_chapel_inline_reduce.py
